writing DAO and class

+++ src/segabank/bo/ComptePayant.java
@@ -6,13 +6,13 @@ public class ComptePayant extends Compte {
     // Si l'utilisateur retire 100€, on lui retirera donc 100*1.05 = 105€ de son compte, soit le cout de chaque opération
     // Si l'utilisateur dépose 100€, on lui ajoutera donc 100*0.95 = 95€ à son compte, les 5€ correspondant à la "taxe"
 
-    private static final double TAUX_OPERATION = 5;
+    private static final Integer TAUX_OPERATION = 5;
 
-    public ComptePayant(double id, double solde) {
+    public ComptePayant(Double id, Double solde) {
         super(id, solde);
     }
 
-    public double getTauxOperation() {
+    public Integer getTauxOperation() {
         return TAUX_OPERATION;
     }
